Update pynus to work silent
Adding parameters to pynus to make it work silent by default.

New arguments:

-a/--address  - set the address of the device to connect to.
-n/--name     - set the name of the device to connect to.
-p/--print    - print a list of available NUS capable devices.
-r/-reconnect - Automatically reconnect to the device upon disconnect.
-l/--logging  - Logging to log file, pynus.py.
-v/--verbose  - Verbose logging to console.

Adding run_pty.sh script to run the pynus.py as uart backend for pty.
This pty instance can subsequently be used as a serial port to
communicate with the device. For example 'rshell' for micropython or
similar.